requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
distlib==0.3.6
Django==4.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-redis==5.2.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.8.0
idna==3.4
image==1.5.33
importlib-metadata==5.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
openapi-codec==1.3.2
Pillow==9.3.0
pipreqs==0.4.13
platformdirs==2.5.2
psycopg2==2.9.4
psycopg2-binary==2.9.4
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2020.1
redis==4.5.1
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
whitenoise==5.1.0
yarg==0.1.9
zipp==3.9.0
